kafka_connectors/requirements.txt
confluent-kafka==2.2.0 
requests==2.31.0
python-dotenv==1.0.0
fastavro==1.8.2
elasticsearch==8.8.2
weaviate-client==3.22.1
opensearch-py==2.2.0
qdrant-client==1.3.1
pinecone-client==2.2.2